Creating model classes for our app
Django model classes describe the structure of our application’s database. And provide us with ways to access it. Each model class corresponds to a database table, and it contains attributes that correspond to the table fields. When we interact with the database we can use class methods to query it and get back class instance objects that correspond to database records. We can also modify the object to alter the database records or create new objects to create new database records.